+++ atuin-client/src/lib.rs
@@ -6,8 +6,6 @@ extern crate log;
 #[cfg(feature = "sync")]
 pub mod api_client;
 #[cfg(feature = "sync")]
-pub mod encryption;
-#[cfg(feature = "sync")]
 pub mod sync;
 
 pub mod database;
